requirements.txt
numpy==1.26.4
protobuf==3.20.3
beautifulsoup4==4.12.3
google==3.0.0 
google-api-core==1.34.1
grpcio==1.64.1 
grpcio-status==1.48.2 
soupsieve==2.5
openai
python-dotenv
fastapi
uvicorn
flask
soundfile
google-generativeai
torch
torchaudio
transformers
scipy
requests
pydantic
pydub
sounddevice
vosk
keyboard